# docs/api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Smarthost API
  version: 1.0.0-draft.8
  summary: Email validation and tracked delivery API (specification 2.10 contract).
  description: |
    Public Smarthost API contract. Normative elaboration of
    `docs/20260908-1644-smarthost-llm-spec.yaml` (specification 2.10), which wins on
    any conflict. A client-facing integration guide is `docs/api/integration-guide.md`
    (served by an installation at `/docs/api`). Status and event values are defined in
    `docs/contracts/status-vocabulary.yaml`.

    ## Conventions
    * The client (tenant) is always derived from the API key. No request carries
      or overrides a `client_id`. Resources owned by another client return `404`.
    * Only `active` and `throttled` clients may create or add work. The
      work-creating operations (create validation job, create send job, add
      recipients, submit) return `403` for `pending_approval` and `suspended`
      clients, which can still read their resources. Keys of `closed` clients
      fail authentication (`401`).
    * Identifiers are UUIDv7 strings. Timestamps are RFC 3339 in UTC (`Z`).
    * JSON field names are `snake_case`. Errors use RFC 9457
      `application/problem+json`.
    * Request bodies are limited to 10 MiB (`413` above it). Collections use
      cursor (keyset) pagination: `limit` and an opaque `cursor`.
    * API keys may expire; an expired key fails exactly like a revoked one (`401`).
...

...
       metadata only.
...

...

...
    URLs.

...
    read.
...
servers:
...
security:
  - apiKey: []
...
paths:
  /validation-jobs:
    post:
      tags: [Validation]
      operationId: createValidationJob
      summary: Create an asynchronous validation job (max 10000 addresses).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ValidationJobCreateRequest'}
      responses:
        '202':
          description: Job accepted (or idempotent replay of an earlier identical request).
          headers:
            Location: {$ref: '#/components/headers/Location'}
            Idempotent-Replayed: {$ref: '#/components/headers/IdempotentReplayed'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ValidationJob'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/IdempotencyInProgress'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
  /validation-jobs/{id}:
    get:
      tags: [Validation]
      operationId: getValidationJob
      summary: Retrieve validation job state and summary.
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Job state.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ValidationJob'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
...
  /send-jobs:
    post:
      tags: [Sending]
      operationId: createSendJob
      summary: Create a send job in `collecting` state (job-level metadata only).
      description: >-
        The sender identity's domain must be a registered sending domain of the
        client. Verification (and, in production, active DKIM) is enforced at
        submit.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SendJobCreateRequest'}
      responses:
        '201':
          description: Job created in `collecting` state (or idempotent replay).
          headers:
            Location: {$ref: '#/components/headers/Location'}
            Idempotent-Replayed: {$ref: '#/components/headers/IdempotentReplayed'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SendJob'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/IdempotencyInProgress'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
...
  /send-jobs/{id}:
    get:
      tags: [Sending]
      operationId: getSendJob
      summary: Retrieve send-job summary.
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Job state.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SendJob'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: opaque API key
      description: >-
        `Authorization: Bearer <api key>`. Keys are shown once at creation and
        stored only as a hash. Revoked keys return 401. API keys are for
        machine clients only and are never accepted for dashboard login.
  headers:
    Location:
      description: URL of the created resource.
      schema: {type: string, format: uri-reference}
    IdempotentReplayed:
      description: >-
        Present with value `true` when the response replays an earlier request with
        the same Idempotency-Key: no side effect was repeated, and the status code,
        Location and resource identity are those of the original operation.
      schema: {type: string, enum: ['true']}
    RetryAfter:
      description: Seconds to wait before retrying.
      schema: {type: integer, minimum: 0}
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Client-generated unique key (a UUID is recommended), scoped to the
        client and to the endpoint (and, for recipient batches, to the send
        job). Same key + same body is a replay: no repeated side effect, the same
        resource, the same status code and Location, and `Idempotent-Replayed:
        true`. For resource creation the body is the resource's current
        representation (response bodies are not stored); a recipient-batch replay
        reproduces its original result, including the running total at that time.
        Same key + different body returns 422. Same key while the first request is
        still being processed returns 409.
      schema: {type: string, minLength: 8, maxLength: 255, pattern: '^[\x21-\x7E]+$'}
    ResourceId:
      name: id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uuid'}
...
  responses:
    BadRequest:
      description: Malformed request (unparseable JSON, bad query parameter, missing Idempotency-Key).
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthorized:
      description: Missing, invalid or revoked API key.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Forbidden:
      description: >-
        The client may not create or add work in its current status
        (`pending_approval` or `suspended`); only `active` and `throttled` clients
        may. Reads remain available.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    NotFound:
      description: Resource does not exist or belongs to another client.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    IdempotencyInProgress:
      description: A request with the same Idempotency-Key is still being processed.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    PayloadTooLarge:
      description: Request body exceeds the configured size limit (10 MiB).
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    UnprocessableEntity:
      description: >-
        Semantically invalid request: schema violation, recipient ceiling
        exceeded, duplicate recipient address, class-specific rules, unknown or
        unverified sending domain, empty job at submit, or an Idempotency-Key
        reused with a different body.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    TooManyRequests:
      description: >-
        Rate limit exceeded (problem type `rate-limited`), or the request would exceed
        one of the client's quotas for the current UTC day or month (problem type
        `quota-exceeded`, with the `quota` member). `Retry-After` gives the seconds
        until a retry can succeed.
      headers:
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
...
    Uuid:
      type: string
      format: uuid
      description: UUIDv7.
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339, UTC.
    ExternalReference:
      type: string
      minLength: 1
      maxLength: 255
      description: Client-supplied opaque reference, scoped to the client. Not required to be unique.
...
    EmailIdentity:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: {type: string, format: email, maxLength: 320}
        name: {type: string, maxLength: 200}
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status]
      properties:
        type: {type: string, format: uri-reference, example: 'https://smarthost.example.com/problems/validation-error'}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        quota:
          description: >-
            Present on `quota-exceeded` (429): the quota the request would have exceeded.
            Only the client's own quota and usage are disclosed.
          type: object
          required: [metric, period, limit, used, requested, resets_at]
          properties:
            metric: {type: string, enum: [validation_jobs, validation_addresses, send_jobs, send_recipients]}
            period: {type: string, enum: [day, month]}
            limit: {type: integer}
            used: {type: integer, description: Admitted in the current period before this request.}
            requested: {type: integer}
            resets_at: {type: string, format: date-time}
        errors:
          type: array
          description: Field-level errors for 422 responses.
          items:
            type: object
            required: [pointer, message]
            properties:
              pointer: {type: string, description: 'JSON Pointer into the request body, e.g. /recipients/3/email_address.'}
              message: {type: string}

    # ---- vocabulary enums (must match docs/contracts/status-vocabulary.yaml)
    ValidationJobStatus:
      type: string
      enum: [queued, processing, completed, failed, cancelled]
    SendJobStatus:
      type: string
      enum: [collecting, queued, processing, dispatched, completed, failed, cancelled]
    MessageClass:
      type: string
      enum: [transactional, subscription]
...
    ValidationAddressInput:
      type: object
      additionalProperties: false
      required: [address]
      properties:
        address:
          type: string
          minLength: 1
          maxLength: 512
          description: Address exactly as held by the client; retained verbatim as `original_address`.
        external_address_reference:
          $ref: '#/components/schemas/ExternalReference'
          description: >-
            Optional opaque client reference for this address (e.g. a source
            record id), separate from the job's external_reference.
    ValidationJobCreateRequest:
      type: object
      additionalProperties: false
      required: [addresses]
      properties:
        external_reference: {$ref: '#/components/schemas/ExternalReference'}
        addresses:
          type: array
          minItems: 1
          maxItems: 10000
          description: Duplicates are permitted and validated individually. Smarthost never rewrites an address.
          items: {$ref: '#/components/schemas/ValidationAddressInput'}
    ValidationJob:
      type: object
      additionalProperties: false
      required: [id, status, submitted_at, total_addresses, processed_count, classification_counts]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        external_reference:
          oneOf:
            - {$ref: '#/components/schemas/ExternalReference'}
            - {type: 'null'}
        status: {$ref: '#/components/schemas/ValidationJobStatus'}
        submitted_at: {$ref: '#/components/schemas/Timestamp'}
        started_at:
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        completed_at:
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        total_addresses: {type: integer, minimum: 1, maximum: 10000}
        processed_count: {type: integer, minimum: 0}
        classification_counts:
          $ref: '#/components/schemas/ClassificationCounts'
    ClassificationCounts:
      type: object
      description: Count of processed addresses per overall_classification. All keys always present.
      additionalProperties: false
      required: [deliverable, probably_deliverable, undeliverable, temporarily_unverifiable, unknown, risky]
      properties:
        deliverable: {type: integer, minimum: 0}
        probably_deliverable: {type: integer, minimum: 0}
        undeliverable: {type: integer, minimum: 0}
        temporarily_unverifiable: {type: integer, minimum: 0}
        unknown: {type: integer, minimum: 0}
        risky: {type: integer, minimum: 0}
...
    TrackingOptions:
      type: object
      additionalProperties: false
      description: Delivery-time instrumentation switches. Both default to false (tracking disabled).
      properties:
        opens:
          type: boolean
          default: false
          description: Insert an open-tracking pixel into the HTML part. Messages without HTML get no open tracking.
        clicks:
          type: boolean
          default: false
          description: >-
            Rewrite eligible absolute HTTP/HTTPS links in the HTML part through
            the click tracker. Never rewritten: plain-text URLs, `mailto:`,
            fragment-only and non-HTTP(S) links, and unsubscribe URLs.
    SendJobCreateRequest:
      type: object
      additionalProperties: false
      required: [external_reference, message_class, sender_identity]
      description: >-
        Job-level metadata only; recipients are added in batches. `client_id`
        and arbitrary headers are not accepted. `list_id` is mandatory for
        `subscription` and not accepted for `transactional`.
      properties:
        external_reference: {$ref: '#/components/schemas/ExternalReference'}
        message_class: {$ref: '#/components/schemas/MessageClass'}
        list_id:
          type: string
          minLength: 1
          maxLength: 255
          description: RFC 2919 `List-Id` header value.
        sender_identity:
          $ref: '#/components/schemas/EmailIdentity'
          description: >-
            Header From. Its domain must be a registered sending domain of the
            client. The envelope return-path is always a Smarthost VERP address.
        reply_to:
          $ref: '#/components/schemas/EmailIdentity'
          description: Optional job-level Reply-To.
        tracking: {$ref: '#/components/schemas/TrackingOptions'}
      if:
        required: [message_class]
        properties: {message_class: {const: subscription}}
      then:
        required: [list_id]
      else:
        not: {required: [list_id]}
...
    SendJob:
      type: object
      additionalProperties: false
      required:
        - id
        - external_reference
        - message_class
        - sender_identity
        - status
        - created_at
        - queued_at
        - started_at
        - dispatch_completed_at
        - completed_at
        - total_recipients
        - summary_counts
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        external_reference: {$ref: '#/components/schemas/ExternalReference'}
        message_class: {$ref: '#/components/schemas/MessageClass'}
        list_id: {type: [string, 'null']}
        sender_identity: {$ref: '#/components/schemas/EmailIdentity'}
        reply_to:
          oneOf: [{$ref: '#/components/schemas/EmailIdentity'}, {type: 'null'}]
        tracking: {$ref: '#/components/schemas/TrackingOptions'}
        status: {$ref: '#/components/schemas/SendJobStatus'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        queued_at:
          description: When the job was sealed by submit.
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        started_at:
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        dispatch_completed_at:
          description: When the job became `dispatched`.
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        completed_at:
          description: When every message reached a terminal knowledge state.
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        total_recipients:
          type: integer
          minimum: 0
          maximum: 10000
          description: Running total while collecting; fixed once sealed.
        summary_counts: {$ref: '#/components/schemas/MessageStatusCounts'}
    MessageStatusCounts:
      type: object
      description: Count of messages per current_status. All keys always present.
      additionalProperties: false
      required:
        - created
        - queued
        - submitted
        - deferred
        - outcome_unknown
        - remote_accepted
        - soft_bounced
        - hard_bounced
        - complained
        - failed
        - suppressed
      properties:
        created: {type: integer, minimum: 0}
        queued: {type: integer, minimum: 0}
        submitted: {type: integer, minimum: 0}
        deferred: {type: integer, minimum: 0}
        outcome_unknown: {type: integer, minimum: 0}
        remote_accepted: {type: integer, minimum: 0}
        soft_bounced: {type: integer, minimum: 0}
        hard_bounced: {type: integer, minimum: 0}
        complained: {type: integer, minimum: 0}
        failed: {type: integer, minimum: 0}
        suppressed: {type: integer, minimum: 0}
...